pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpios-setup"
version = "0.1.0"
description = "Declarative Raspberry Pi OS setup utility (apt, Pi-Apps, desktop, services)"
authors = [{name = "You"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12.3",
  "pyyaml>=6.0.2",
  "jinja2>=3.1.4",
]

[project.scripts]
rpios-setup = "rpios_setup.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["rpios_setup*"]
exclude = ["configs*", "templates*", "tests*", "scripts*"]
